refactor: nombres de carpeta y useCase
