+++ app/src/debug/java/com/app/debugGrader/ui/pages/config/DebugConfigViewModel.kt
@@ -5,7 +5,7 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.app.grader.data.database.AppDatabase
 import com.app.debugGrader.data.database.DebugDatabaseSeeder
-import com.app.grader.infrastructure.review.InAppReviewHelper
+import com.app.grader.service.review.InAppReviewHelper
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
